feat(api/student): make a separate handler type file for students
