+++ apps/api/src/handlers/student/get_student.rs
@@ -2,7 +2,7 @@ use axum::Json;
 use axum::extract::{Query, State};
 
 use crate::domain::models::student::StudentError;
-use crate::handlers::student::{ListStudentResponse, StudentResponse, to_response};
+use crate::handlers::student::types::{ListStudentResponse, StudentResponse, to_response};
 use crate::infra::errors::InfraError;
 use crate::infra::repositories::student_repository::{
     StudentFilter, get, get_with_filter, get_with_roll_no,
